requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.3.1=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.9.9=haa95532_0
certifi=2025.10.5=pypi_0
charset-normalizer=3.4.4=pypi_0
colorama=0.4.6=pypi_0
expat=2.7.1=h8ddb27b_0
filelock=3.19.1=pypi_0
fsspec=2025.9.0=pypi_0
grpcio=1.76.0=pypi_0
huggingface-hub=0.36.0=pypi_0
idna=3.11=pypi_0
jinja2=3.1.6=pypi_0
libffi=3.4.4=hd77b12b_1
libzlib=1.3.1=h02ab6af_0
markdown=3.9=pypi_0
markupsafe=2.1.5=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.3=pypi_0
numpy=2.1.2=pypi_0
openssl=3.0.18=h543e019_0
packaging=25.0=pypi_0
pandas=2.3.3=pypi_0
pillow=11.3.0=pypi_0
pip=25.2=pyhc872135_1
protobuf=6.33.0=pypi_0
python=3.10.19=h981015d_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.3=pypi_0
regex=2025.10.23=pypi_0
requests=2.32.5=pypi_0
safetensors=0.6.2=pypi_0
setuptools=80.9.0=py310haa95532_0
six=1.17.0=pypi_0
sqlite=3.50.2=hda9a48d_1
sympy=1.14.0=pypi_0
tensorboard=2.20.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tk=8.6.15=hf199647_0
tokenizers=0.22.1=pypi_0
torch=2.7.1+cu118=pypi_0
torchaudio=2.7.1+cu118=pypi_0
torchvision=0.22.1+cu118=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.57.1=pypi_0
typing-extensions=4.15.0=pypi_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.5.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310haa95532_0
xz=5.6.4=h4754444_1
zlib=1.3.1=h02ab6af_0
